+++ internal/cache/in_memory_cache_test.go
@@ -1,7 +1,7 @@
 package cache_test
 
 import (
-	"github.com/thomaspoignant/go-feature-flag/dto"
+	"github.com/thomaspoignant/go-feature-flag/model/dto"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
